// app/simulation/step.enum.ts
export enum StepEnum{
    STEP_CLIENT = 0,
    STEP_VEHICLE = 1,
    STEP_SIMULATION = 2,
    STEP_CUSTOMER_CARD = 3,
    STEP_SEND = 4,
    STEP_DOCUMENT = 5,
    STEP_MOBILE_SUBMISSION = 6,
    STEP_MOBILE_CLIENT_DATA = 7,
    STEP_MOBILE_ADDRESS_DATA = 8,
    STEP_MOBILE_SPOUSE = 9,
    STEP_MOBILE_PROFESSIONAL_DATA = 10,
    STEP_MOBILE_PERSON_DATA = 11,
    STEP_MOBILE_BANKING_REFENCES = 12,
    STEP_MOBILE_VEHICLE_DATA = 13,
    STEP_MOBILE_GUARANTOR_ONE_CLIENT = 14,
    STEP_MOBILE_GUARANTOR_ONE_ADDRESS = 15,
    STEP_MOBILE_GUARANTOR_ONE_SPOUSE = 16,
    STEP_MOBILE_GUARANTOR_ONE_PROFESSIONAL = 17,
    STEP_MOBILE_GUARANTOR_ONE_PERSON = 18,
    STEP_MOBILE_GUARANTOR_ONE_BANKING = 19,
    STEP_MOBILE_GUARANTOR_TWO_CLIENT = 20,
    STEP_MOBILE_GUARANTOR_TWO_ADDRESS = 21,
    STEP_MOBILE_GUARANTOR_TWO_SPOUSE = 22,
    STEP_MOBILE_GUARANTOR_TWO_PROFESSIONAL = 23,
    STEP_MOBILE_GUARANTOR_TWO_PERSON = 24,
    STEP_MOBILE_GUARANTOR_TWO_BANKING = 25,
}